+++ test/JSON2CSVTransform.js
@@ -13,7 +13,7 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
     const opts = {
       fields: ['carModel', 'price', 'color', 'transmission']
     };
-    const transformOpts = { readableObjectMode: true, writableObjectMode: true };
+    const transformOpts = { objectMode: true };
 
     const transform = new Json2csvTransform(opts, transformOpts);
     const processor = input.pipe(transform);
@@ -25,7 +25,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.ndjson);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should handle ndjson', (t) => {
@@ -44,7 +47,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.ndjson);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should error on invalid ndjson input data', (t) => {
@@ -61,7 +67,7 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
       t.end();
     });
     processor.on('error', (err) => {
-      t.ok(err.message.indexOf('Invalid JSON') !== -1);
+      t.ok(err.message.includes('Invalid JSON'));
       t.end();
     });
   });
@@ -80,7 +86,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.deepEqual(opts, {});
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should error if input data is not an object', (t) => {
@@ -115,7 +124,7 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
       t.end();
     });
     processor.on('error', (err) => {
-      t.ok(err.message.indexOf('Invalid JSON') !== -1);
+      t.ok(err.message.includes('Invalid JSON'));
       t.end();
     });
   });
@@ -135,7 +144,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.emptyObject);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should handle empty array', (t) => {
@@ -153,7 +165,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.emptyObject);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should hanlde array with nulls', (t) => {
@@ -175,7 +190,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.emptyObject);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should handle deep JSON objects', (t) => {
@@ -189,7 +207,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.deepJSON);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should parse json to csv and infer the fields automatically ', (t) => {
@@ -204,7 +225,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.defaultStream);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should parse json to csv using custom fields', (t) => {
@@ -223,7 +247,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.end();
   
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should output only selected fields', (t) => {
@@ -241,7 +268,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.selected);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should output fields in the order provided', (t) => {
@@ -259,7 +289,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.reversed);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should output empty value for non-existing fields', (t) => {
@@ -277,7 +310,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.withNotExistField);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should name columns as specified in \'fields\' property', (t) => {
@@ -301,7 +337,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.fieldNames);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
 
@@ -371,7 +410,38 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.nested);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
+  });
+
+  testRunner.add('field.value function should receive a valid field object', (t) => {
+    const opts = {
+      fields: [{
+        label: 'Value1',
+        default: 'default value',
+        value: (row, field) => {
+          t.deepEqual(field, { label: 'Value1', default: 'default value' });
+          return row.value1.toLocaleString();
+        }
+      }]
+    };
+
+    const transform = new Json2csvTransform(opts);
+    const processor = jsonFixtures.functionStringifyByDefault().pipe(transform);
+
+    let csv = '';
+    processor
+      .on('data', chunk => (csv += chunk.toString()))
+      .on('end', () => {
+        t.equal(csv, csvFixtures.functionStringifyByDefault);
+        t.end();
+      })
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('field.value function should stringify results by default', (t) => {
@@ -392,7 +462,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.functionStringifyByDefault);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('field.value function should not stringify if stringify is selected to false', (t) => {
@@ -414,7 +487,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.functionNoStringify);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should process different combinations in fields option', (t) => {
@@ -448,7 +524,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.fancyfields);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // Preprocessing
@@ -469,13 +548,16 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.unwind);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should support multi-level unwind', (t) => {
     const opts = {
-      fields: ['carModel', 'price', 'items.name', 'items.color', 'items.items.position', 'items.items.color'],
-      unwind: ['items', 'items.items']
+      fields: ['carModel', 'price', 'extras.items.name', 'extras.items.color', 'extras.items.items.position', 'extras.items.items.color'],
+      unwind: ['extras.items', 'extras.items.items']
     };
 
     const transform = new Json2csvTransform(opts);
@@ -488,15 +570,18 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.unwind2);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
 
 
   testRunner.add('should unwind and blank out repeated data', (t) => {
     const opts = {
-      fields: ['carModel', 'price', 'items.name', 'items.color', 'items.items.position', 'items.items.color'],
-      unwind: ['items', 'items.items'],
+      fields: ['carModel', 'price', 'extras.items.name', 'extras.items.color', 'extras.items.items.position', 'extras.items.items.color'],
+      unwind: ['extras.items', 'extras.items.items'],
       unwindBlank: true
     };
 
@@ -510,7 +595,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.unwind2Blank);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should support flattening deep JSON', (t) => {
@@ -528,7 +616,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.flattenedDeepJSON);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should support custom flatten separator', (t) => {
@@ -547,7 +638,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.flattenedCustomSeparatorDeepJSON);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should unwind and flatten an object in the right order', (t) => {
@@ -566,7 +660,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.unwindAndFlatten);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // Default value
@@ -587,7 +684,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.defaultValueEmpty);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should override \'options.defaultValue\' with \'field.defaultValue\'', (t) => {
@@ -610,7 +710,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.overriddenDefaultValue);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should use \'options.defaultValue\' when no \'field.defaultValue\'', (t) => {
@@ -642,7 +745,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.overriddenDefaultValue);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // Quote
@@ -663,7 +769,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.withSimpleQuotes);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should be able to don\'t output quotes when setting \'quote\' to empty string', (t) => {
@@ -682,7 +791,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.withoutQuotes);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should escape quotes when setting \'quote\' property is present', (t) => {
@@ -701,7 +813,31 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.escapeCustomQuotes);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
+  });
+
+  testRunner.add('should not escape \'"\' when setting \'quote\' set to something else', (t) => {
+    const opts = {
+      quote: '\''
+    };
+
+    const transform = new Json2csvTransform(opts);
+    const processor = jsonFixtures.doubleQuotes().pipe(transform);
+
+    let csv = '';
+    processor
+      .on('data', chunk => (csv += chunk.toString()))
+      .on('end', () => {
+        t.equal(csv, csvFixtures.doubleQuotesUnescaped);
+        t.end();
+      })
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // Double Quote
@@ -717,10 +853,13 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.quotes);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
-  testRunner.add('should not escape quotes with double quotes, when there is a backslah in the end', (t) => {
+  testRunner.add('should not escape quotes with double quotes, when there is a backslash in the end', (t) => {
     const transform = new Json2csvTransform();
     const processor = jsonFixtures.backslashAtEnd().pipe(transform);
 
@@ -731,10 +870,13 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.backslashAtEnd);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
-  testRunner.add('should not escape quotes with double quotes, when there is a backslah in the end, and its not the last column', (t) => {
+  testRunner.add('should not escape quotes with double quotes, when there is a backslash in the end, and its not the last column', (t) => {
     const transform = new Json2csvTransform();
     const processor = jsonFixtures.backslashAtEndInMiddleColumn().pipe(transform);
 
@@ -745,7 +887,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.backslashAtEndInMiddleColumn);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should escape quotes with value in \'doubleQuote\'', (t) => {
@@ -764,7 +909,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.doubleQuotes);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should escape quotes before new line with value in \'doubleQuote\'', (t) => {
@@ -782,7 +930,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.backslashBeforeNewLine);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // Delimiter
@@ -803,7 +954,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.tsv);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should remove last delimiter |@|', (t) => {
@@ -819,7 +973,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.delimiter);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // EOL
@@ -840,7 +997,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.eol);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // Excell
@@ -861,7 +1021,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.excelStrings);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // Escaping and preserving values
@@ -881,7 +1044,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.trailingBackslash);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should escape " when preceeded by \\', (t) => {
@@ -895,7 +1061,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.escapeDoubleBackslashedDoubleQuote);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should preserve new lines in values', (t) => {
@@ -912,12 +1081,15 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
       .on('end', () => {
         t.equal(csv, [
       '"a string"',
-      '"with a \ndescription\\n and\na new line"',
-      '"with a \r\ndescription and\r\nanother new line"'
+      '"with a \u2028description\\n and\na new line"',
+      '"with a \u2029\u2028description and\r\nanother new line"'
     ].join('\r\n'));
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should preserve tabs in values', (t) => {
@@ -931,7 +1103,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.escapeTab);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // Header
@@ -952,7 +1127,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.withoutHeader);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // Include empty rows
@@ -968,7 +1146,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.emptyRowNotIncluded);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should include empty rows when options.includeEmptyRows is true', (t) => {
@@ -986,7 +1167,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.emptyRow);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should not include empty rows when options.includeEmptyRows is false', (t) => {
@@ -1004,7 +1188,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.emptyRowNotIncluded);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should include empty rows when options.includeEmptyRows is true, with default values', (t) => {
@@ -1035,7 +1222,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.emptyRowDefaultValues);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   testRunner.add('should parse data:[null] to csv with only column title, despite options.includeEmptyRows', (t) => {
@@ -1059,7 +1249,10 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv, csvFixtures.emptyObject);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 
   // BOM
@@ -1083,6 +1276,9 @@ module.exports = (testRunner, jsonFixtures, csvFixtures, inMemoryJsonFixtures) =
         t.equal(csv.length, csvFixtures.withBOM.length);
         t.end();
       })
-      .on('error', err => t.notOk(true, err.message));
+      .on('error', err => {
+        t.notOk(true, err.message)
+        t.end();  
+      });
   });
 };